health checks folder changed to use dashes to split names. require call adjusted to point to this folder
